refactor: 💡 move KbnTopNavControllerProvider into console
