+++ src/legacy/core_plugins/console/public/src/controllers/__tests__/kbn_top_nav_controller.js
